# Keccak-F1600: Use d-form LDP/LDR instead of LD2