The 1-fold Keccak-F1600 implementation based on v8.4-A Neon instructions uses a 0-lane LD2/ST2 to load/store the Keccak-F1600 state from memory. The upper halves of the Neon registers are unused.

This commit simplifies this by instead using a LDP/LDR targeting the D-forms of the Neon registers.

This matches the version verified by @jargh