@hanno-becker 2025-02-26
Benchmarks on my local M1 -- basically no change, as expected:

Before:
```
c889f3ee3e79:mlkem-native beckphan$ tests bench -c M1 -r --opt=OPT
INFO  > Benchmark          Compile     (native opt):     make bench OPT=1 AUTO=1 CYCLES=M1 -j10
INFO  > Benchmark          ML-KEM-512  (native opt):     EXEC_WRAPPER=sudo make run_bench_512
INFO  > Benchmark          ML-KEM-512  (native opt):     sudo test/build/mlkem512/bin/bench_mlkem512
   keypair cycles = 12506
    encaps cycles = 15477
    decaps cycles = 20275

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  12478  12489  12496  12499  12502  12506  12510  12513  12517  12526  13588
    encaps percentiles:  15450  15461  15467  15471  15474  15477  15480  15484  15489  15495  16672
    decaps percentiles:  20242  20257  20261  20266  20270  20275  20284  20294  20305  20323  21350

INFO  > Benchmark          ML-KEM-768  (native opt):     EXEC_WRAPPER=sudo make run_bench_768
INFO  > Benchmark          ML-KEM-768  (native opt):     sudo test/build/mlkem768/bin/bench_mlkem768
   keypair cycles = 22030
    encaps cycles = 24679
    decaps cycles = 31689

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  21973  21996  22008  22016  22022  22030  22041  22049  22060  22100  23154
    encaps percentiles:  24618  24644  24656  24663  24670  24679  24686  24696  24710  24747  25791
    decaps percentiles:  31626  31657  31669  31678  31683  31689  31698  31706  31717  31748  32807

INFO  > Benchmark          ML-KEM-1024 (native opt):     EXEC_WRAPPER=sudo make run_bench_1024
INFO  > Benchmark          ML-KEM-1024 (native opt):     sudo test/build/mlkem1024/bin/bench_mlkem1024
   keypair cycles = 31201
    encaps cycles = 36386
    decaps cycles = 45000

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  31104  31153  31169  31181  31192  31201  31212  31232  31257  32244  33295
    encaps percentiles:  36269  36311  36335  36356  36368  36386  36407  36432  36460  37484  37622
    decaps percentiles:  44907  44952  44965  44977  44990  45000  45015  45028  45057  46040  46200

All good!```
After:
```
=====
   keypair cycles = 12481
    encaps cycles = 15468
    decaps cycles = 20254

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  12455  12466  12471  12474  12478  12481  12485  12489  12494  12501  13565
    encaps percentiles:  15438  15453  15457  15461  15465  15468  15471  15476  15479  15486  16668
    decaps percentiles:  20224  20234  20239  20244  20248  20254  20260  20273  20285  20301  21337

INFO  > Benchmark          ML-KEM-768  (native opt):     EXEC_WRAPPER=sudo make run_bench_768
INFO  > Benchmark          ML-KEM-768  (native opt):     sudo test/build/mlkem768/bin/bench_mlkem768
   keypair cycles = 21994
    encaps cycles = 24652
    decaps cycles = 31647

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  21938  21961  21973  21979  21986  21994  22003  22013  22022  22062  23093
    encaps percentiles:  24592  24615  24627  24636  24644  24652  24659  24666  24679  24711  25754
    decaps percentiles:  31584  31616  31626  31635  31641  31647  31656  31664  31672  31705  32758

INFO  > Benchmark          ML-KEM-1024 (native opt):     EXEC_WRAPPER=sudo make run_bench_1024
INFO  > Benchmark          ML-KEM-1024 (native opt):     sudo test/build/mlkem1024/bin/bench_mlkem1024
   keypair cycles = 31176
    encaps cycles = 36377
    decaps cycles = 44949

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  31088  31128  31139  31152  31163  31176  31190  31206  31231  32220  33288
    encaps percentiles:  36237  36292  36319  36340  36360  36377  36396  36417  36446  37483  37778
    decaps percentiles:  44852  44896  44912  44923  44935  44949  44960  44979  45003  45997  46127

All good!
```