@hanno-becker 2025-02-26
I double-checked in https://github.com/pq-code-package/mlkem-native/actions/runs/13537777600 that we do actually run this code in CI.